fix: making not a header-only library
